requirements.txt
tensorflow
numpy >= 1.18.1
matplotlib
gym
pygame
gitpython
tensorflow-probability
